ScheduleItem table is now just a seed to fill in the database
This commit removes the use of the ScheduleItem as the source of the
schedule (for managing and updating the database; the web server has
already been changed). The ScheduleItem table is now only used to indicate
to the Obscat executable that there are new items to access.

There have been some code updates and a compiler restriction, to ghc 7.8
or greater (effectively, due to the base constraint). aeson version 0.10
is excluded, and 0.11 is now allowed (but not yet tested).