Improve handling of non-science obs queries
This changeset attempts to support both "discarded" and "data-not-from-obscat"
non-science observations. It is complicated by being forced to stuff logic
into the nsName field rather than by adding extra fields to the data type;
this was done to avoid upgrading the database schema.

I am not 100% convinced that this change is sensible, in that there now
seem to be a lot of "discarded" non-science observations. This may just
be my mis-understanding of the ObsCat results.

At some level, it may not matter, because the non-science observations
may only be of interest for the live / running-now schedule, and not
historically.