Change visual display of constellation list
Style the table so that there can be multiple rows on as screen line,
as the table is ling, and there's only 2 elements per row.

I also experimented with adding a sky view to this page, with the
idea being that the constellations would be colored based on the
exposure time. This is commented out since the current code does not
fill these outlines "well" (due to determining which part of the sky
is inside vs outside an outline).