Make rows appear to be selected
This uses shadows to highlight the currently-selected row of the
table (which has been converted to a bunch of separate blocks). It
also attempts to make the whole tr element be selectable, but it
doesn't appear to be possible with the current HTML output. Or at
least, not easily and cleanly. The current approach is a balance,
but may end up being confusing. Let's see.