Hopefully fix running the webserver on Heroku
It appears that the move from a Simple build to a Configure build - i.e.
Setup.hs has changed to support configuring several files during the build
- breaks the 'cabal run webserver' command when using cabal-install 1.18
(version range uncertain), which is what is used on Heroku. The error
message is "cabal: Unrecognised flags: exe:webserver".

Rather than try and fix this, which looks like having to delve deep into
the Cabal library, I have just hacked the run command, assuming that the
paths are correct. An alternative would be to move to a newer cabal/ghc
combination on Heroku, but that's also more effort!