Move the nav bar into the main section.
The main change is actually to support the experimental js view of the
data, which moves the HTML generation back to the server rather than
having it in the JS code.

The new version of the JS code doesn't require the proxy code for the
DSS/ROSAT images, but it hasn't yet been removed.