Added a Makefile mainly to document some of the management steps
This Makefile is intended to simplify management steps - e.g. build
and push the Docker images - rather than to support dependency resolution
(so it won't avoid excess work or know what sub-goals are also needed).
It's main aim is to have a single place to document some of the steps
needed to manage the chandraobs site.