Rework the data ingest and clean up the database
The data ingest has been cleaneed up, *but* the documentation in notes/
has not been updated. It is also incomplete, as all that has been done is
the equivalent of the initdb + getschedule step. What is needed is the
addition of an upateschedule tool, to update ObsIds that are not
archived/finished.

The database structure has been updated to add extra tables - e.g.
tracking so-called invalid obsids - as well as to better match the
existing data structures (e.g. making start time be optional).