Added updateschedule
Working through issues with processing the short-term schedule.

One change is that RA and Dec are now considered equal within a
tolerance, because it seems that when serializing the value to and
from the database small changes could be seen. The tolerance used
is a lot larger than the differences seen due to the serialization,
but is still much smaller than is meaningful (i.e. in the micro-arcsecond
range).