Reduce time to create a timeline.
Applied the same approach to creating the timeline as I used for the
schedules which (on my local test machine) reduces the time taken to
respond to a request for data from ~ 10s to ~2s. I am applying this now
rather than cleaning up the code, as I want this fix live.

There's also a few minor unrelated hlint (and manual) code cleanups.