Improve (?) information when JS is turned off.
Attempt to provide more text indicating that "something should be here,
but it requires JavaScript and I am too lazy to get it to work when
JavaScript is not enabled".

I am not convinced this is the best approach (the schedule page has
a paragraph, but it is likely that - given the potentially-long data-load
times for some schedules - there will be confusion/visual flashing).