Add cache for WWT/Simbad mappings/exp time breakdown
Expand the use of cache for JSON data from the WWT FOV case to the
SIMBAD mappings and exposure-time breakdowns.

It would be good to also add this to the API timeline data, but
there are difficulties (since the live time for this data depends
on when the next data set goes public - and possibly others).