QuerySimbad: ignore terms after '['
A number of CAL targets are of the form "xxx [instrument config]" and
so just remove any "[..." from the search term. There are several
which have "valid" text that this removes, but so far they don't get
resolved in SIMBAD either way, so leave this simple scheme for now.

The scheme probably removes some of the other checks, but leave in for
now.