Test: add basic timeline information
This just adds the previous, current, and next obsids to the top of the
screen. There is some clean up in the code so that the display is
correctly scrolled when an observation is selected.

Code removal and clean up (e.g. prefer addEventListener rather than
setting the onXXX attribute). I have noticed some odd behavior with the
select menu for choosing the background, and I wonder if it could be due
to these changes (e.g. a signal is now propogating when it used to be
eaten up).