Test: change how timeline box works
The timeline box now displays the selected (by user) observation and its
preceeding/following observations, and uses names rather than obsids.

The display is not updated when the selected observation is closed, which
it probably should be.

There is also an issue in the rendering of the timeline, since the contents
are now being updated, which can lead to distorted text/visuals. I am not
sure how to "fix" this.