Test: ensure status pane has a consistent height
Previously the height of the pane would vary depending on which tab was
selected, but it makes more sense to enforce a given height for science
obs (since we know that the details tab is going to be tall).

Would it be useful to allow the pane's height to be adjusted by the user?