Update Dockerfile
There seems to be an issue with updating imagemagick in the
currently-available Ubuntu 18 repository (there's a conflict
over a configuration file which makes the automated update
fail), so I have chosen to remove the offending package
before updating anything.

There appears to have been a change in stack behavior, which
means we need to copy over the cabal file before setting up
stack.