QuerySimbad - change name check
Since the target names have been changed the previous attempt to
find "sensible" names from the target name are no-longer
meaningful, so apply several minor fixes.

There is an issue (which was in the old code too) of what happens
if we end up creating a search term which we already know about.
In this case we end up in the "errr, need to delete something"
check when we should just be linking to an existing field.