Add autocomplete support to test name field
We now add target-name completion to the name field for the
test interface. To avoid un-needed searches the name search now
returns the RA/Dec of each name (randomly-selected from the
possible list of matches) and this is used to trigger the
WWT zoom.

The existing name search has been updated to use this new
endpoint.