webserver: timeline can now be used to select observations
This is smoewhat of a hack, since we take the SVG output
created by vegaEmbed and then go through each of the relevant
observations and turn them into "clickable links" which will
select the ObsId. So this relies on

- the format / labelling used by vegeEmbed for the path
  elements (to select just the ones we care about)

- the storage of the actual data elements in the SVG via the
  __data__.datum dictionary

I had tried using the VegaLite hyperlink support (the "href"
encoding) but we actually don't want a HREF here; we just want to
have a way of calling a JavaScript function to show an ObsId.