Timeline: fix display of the "now" line
From https://vega.github.io/vega-lite/examples/layer_falkensee.html
I decided to actually treat the line as a separate set of data,
rather than just send in a value via the encoding channels.

This fixes up the size and opacity settings I had been struggling
with.

It also makes it clearer (to me) that the data is being drawn at
the current timezone rather than UTC - this is not ideal but at
least it is now internally consistent.

Todo: work out how to get it to display with UTC values for
the X axis.