webserver: add a timeline-view for related observations
This doesn't work as well as the schedule because

a) we tend to use the same instrument so the "faceting" of the
   display is not helpful

   I'd like to sub-facet on SIMBAD class but that requires sending
   in more details and the Vega-Lite logic is unclear.

b) we don't have a "timeline" since we use points as the exposure times
   are generally too small

c) the x axis doesn't get displayed well in many cases

The logic for creating/showing the "View related observations" is
rather complex as, unlike "View Timeline" we do not have the button
in all pages and the techniques used to clean up old panes has left
the brower all-kind-of messed up over what exist or not.

The current-time is now drawn with a tooltip.