Timeline: zoom and tooltip
We can now zoom into the timelines (and reset with a double click)
but we do not tell the users this. I want to explore the approach
a bit more before adding documentation about it.

The tooltips have been changed so that instead of creating a single
label with the information we just include the fields we care about.