Timeline: tweak display
For the related view, switch to showing whether the data is public
or not, not whether it has been "processed".

For the timeline view, add an overview (so no splitting on
instrument) and try to improve the interactive nature of zooming
in. It's surprisingly tricky, and somehow I have "lost" the
legend opacity changes.