Improve links: SIMBAD and coordinates
SIMBAD - in mid 2022 - changed the location of the object-type
information, so use the new location. It's a "WEB 2.0" style
page so you can't seem to control the behaviour via the URL,
which is great....

I need to review the changes to the SIMBAD hierarchy, as it looks
to have changed, but that's for another day.

Chris Mihos's coordinate page has gone the way of the dodo, so
remove it.